backend/requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.2.31
dnspython==2.7.0
PyJWT==2.10.1
pymongo==4.10.1
sqlparse==0.5.3
tzdata==2024.2
